Fix type correctness: use proper types
Change variable and function types to match their actual usage, eliminating
unnecessary casts and fixing sign conversion warnings:

- Change tick counter from int to uint32_t (loaded from network as uint32_t,
  only incremented, never negative)
- Update related functions: sdl_pre_do, set_map_values, prefetch_game,
  prefetch(), next_tick() to use uint32_t for tick parameters/returns
- Change buffer size variables (ticksize, indone, lastticksize) from int
  to size_t (represent buffer positions/sizes)
- Change inused and outused from int to size_t (buffer usage counters)
- Change client_send parameter from int len to size_t len
- Change socktimeout from int to time_t (stores time() return value)
- Make decrypt, send_info, and auto_tick static (only used within file)
- Fix level2exp to cast pow() result (pow returns double, function returns int)
- Increase MAXPASSWORD from 16 to 17 to accommodate null terminator

This addresses compiler warnings about implicit sign conversions and missing
prototypes by fixing the root cause (incorrect types) rather than adding casts.

+++ src/amod/amod.h
@@ -70,7 +70,7 @@ DLL_IMPORT int exp2level(int val);
 DLL_IMPORT int level2exp(int level);
 DLL_IMPORT int mil_rank(int exp);
 // client / server communication
-DLL_IMPORT void client_send(void *buf, int len);
+DLL_IMPORT void client_send(void *buf, size_t len);
 
 
 // ---------- Client exported data structures -------------
